Verify volume's state before attaching it in the cinder_local driver
We want to make sure that the volume isn't already attached somewhere
else before trying to attach it. In the case where it's already attached
locally, simply return silently. It also verifies that the volume is
attachable by checking its status, that should be 'available'.

Also cleanup the persisted secret on failure.